lab8/requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohttp==3.9.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colored==2.3.1
contourpy==1.3.3
cycler==0.12.1
Faker==37.12.0
Flask==3.1.2
fonttools==4.60.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
magic-filter==1.0.12
MarkupSafe==3.0.3
matplotlib==3.10.7
multidict==6.7.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
pydantic==2.5.3
pydantic_core==2.14.6
pyjokes==0.8.3
pyparsing==3.2.5
pyTelegramBotAPI==4.29.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==22.5
pytz==2025.2
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
termcolor==3.2.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.22.0
